Add activiti-spring-cache-manager cache provider
