+++ activiti-cloud-service-common/activiti-cloud-services-common-identity-keycloak/src/main/resources/keycloak-client.properties
@@ -7,3 +7,7 @@ spring.security.oauth2.client.registration.keycloak.scope=openid
 
 spring.security.oauth2.client.provider.keycloak.authorization-uri=${keycloak.auth-server-url}/realms/${keycloak.realm}/protocol/openid-connect/authorize
 spring.security.oauth2.client.provider.keycloak.token-uri=${keycloak.auth-server-url}/realms/${keycloak.realm}/protocol/openid-connect/token
+
+activiti.spring.cache-manager.caches.userRoleMapping.caffeine.spec=maximumSize=${identity.client.cache.cacheMaxSize:1000}, expireAfterWrite=${identity.client.cache.cacheExpireAfterWrite:PT5m}, recordStats
+activiti.spring.cache-manager.caches.userGroups.caffeine.spec=maximumSize=${identity.client.cache.cacheMaxSize:1000}, expireAfterWrite=${identity.client.cache.cacheExpireAfterWrite:PT5m}, recordStats
+activiti.spring.cache-manager.caches.groupRoleMapping.caffeine.spec=maximumSize=${identity.client.cache.cacheMaxSize:1000}, expireAfterWrite=${identity.client.cache.cacheExpireAfterWrite:PT5m}, recordStats
