+++ activiti-cloud-service-common/activiti-cloud-services-common-security/pom.xml
@@ -21,8 +21,8 @@
       <artifactId>activiti-api-runtime-shared</artifactId>
     </dependency>
     <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-cache</artifactId>
+      <groupId>org.activiti</groupId>
+      <artifactId>activiti-spring-cache-manager</artifactId>
     </dependency>
     <dependency>
       <groupId>com.github.ben-manes.caffeine</groupId>
