+++ activiti-cloud-service-common/activiti-cloud-services-common-security/src/main/java/org/activiti/cloud/services/common/security/config/CommonSecurityAutoConfiguration.java
@@ -15,10 +15,8 @@
  */
 package org.activiti.cloud.services.common.security.config;
 
-import com.github.benmanes.caffeine.cache.Caffeine;
 import feign.RequestInterceptor;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.List;
 import org.activiti.api.runtime.shared.security.PrincipalGroupsProvider;
 import org.activiti.api.runtime.shared.security.PrincipalIdentityProvider;
@@ -48,12 +46,9 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
-import org.springframework.cache.Cache;
-import org.springframework.cache.CacheManager;
-import org.springframework.cache.caffeine.CaffeineCache;
-import org.springframework.cache.support.SimpleCacheManager;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
 import org.springframework.core.convert.converter.Converter;
@@ -74,6 +69,7 @@ import org.springframework.web.cors.CorsConfiguration;
 @ConditionalOnWebApplication
 @ConditionalOnMissingBean(value = { SessionAuthenticationStrategy.class, SessionAuthenticationStrategy.class })
 @Import(CommonJwtAuthenticationConverterConfiguration.class)
+@PropertySource("classpath:config/activiti-cloud-services-common-security.properties")
 public class CommonSecurityAutoConfiguration {
 
     private final AuthorizationConfigurer authorizationConfigurer;
@@ -234,18 +230,5 @@ public class CommonSecurityAutoConfiguration {
 
         return request ->
             actuatorMatcher.matches(request) && excludeMatchers.stream().noneMatch(matcher -> matcher.matches(request));
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    public CacheManager cacheManager(Collection<Cache> caches) {
-        SimpleCacheManager cacheManager = new SimpleCacheManager() {
-            @Override
-            protected Cache getMissingCache(String name) {
-                return new CaffeineCache(name, Caffeine.newBuilder().build());
-            }
-        };
-        cacheManager.setCaches(caches);
-        return cacheManager;
     }
 }
