+++ activiti-cloud-service-common/activiti-cloud-services-common-security/src/main/resources/config/activiti-cloud-services-common-security.properties
@@ -1,0 +1,3 @@
+activiti.spring.cache-manager.caches.userSearch.caffeine.spec=maximumSize=${identity.client.cache.cacheMaxSize:1000}, expireAfterWrite=${identity.client.cache.cacheExpireAfterWrite:PT5m}, recordStats
+activiti.spring.cache-manager.caches.groupSearch.caffeine.spec=maximumSize=${identity.client.cache.cacheMaxSize:1000}, expireAfterWrite=${identity.client.cache.cacheExpireAfterWrite:PT5m}, recordStats
+activiti.spring.cache-manager.caches.userInfoApiCall.caffeine.spec=maximumSize=${identity.client.cache.cacheMaxSize:1000}, expireAfterWrite=${identity.client.cache.cacheExpireAfterWrite:PT5m}, recordStats
